Fix layout and add chapter icons
