+++ src/components/ComponentFactory.js
@@ -6,8 +6,8 @@ import BlockQuote from './BlockQuote';
 import Button from './Button';
 import Card from './Card';
 import CardGroup from './CardGroup';
-import Chapter from './Chapter';
-import Chapters from './Chapters';
+import Chapter from './Chapters/Chapter';
+import Chapters from './Chapters/Chapters';
 import Code from './Code';
 import Cond from './Cond';
 import Container from './Container';
